Move test certs and pem cert pool to examples/ct
Cert pool depends on CT specific versions of x.509 and test certs are CT
specific so should be kept with that code.
